Fix bug in benchmark agent
This was first detected in the static zoning agent. Calls in the
higher floors of a zone were unserved for long times, pushing up
the waiting/system time metrics. The reason this happened was that
elevators, while en route to the queued floor with the longest
waiting time, picked up other passengers, causing them to abandon
their original pick-up route in favor of delivering the newly
boarded passenger. This can be (and was) fixed by triggering a
button press in the `handle_waiting` function in the
`BenchmarkAgent` implementation.